[EUREKA-561-4]. Add validateUserUnaffiliatedLocationUpdates method to Order Open/unopen (#1061)
* [EUREKA-561-3]. Add validateUserUnaffiliatedLocationUpdates method to Order Open/unopen

* [EUREKA-561-4]. Change method args & remove risky conversions

* [EUREKA-561-4]. Replace filtering with StringUtils.equals(), change args names

(cherry picked from commit 8092d9730777bbeacabcf502323235fcbcd92b71)
